refactor: features_github_repository -> samples_github_repository
