+++ apps/app/lib/main_page.dart
@@ -1,8 +1,8 @@
-import 'package:features_github_repository/data.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app/gen/l10n/l10n.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import 'package:samples_github_repository/data.dart';
 
 class MainPage extends ConsumerWidget {
   const MainPage({
