--- packages/features/github_repository/README.md
@@ -1,1 +1,0 @@
-# features_github_repository package
