+++ packages/samples/github_repository/.gitignore
