+++ packages/samples/github_repository/lib/src/domain/model/github_repository.freezed.dart
