+++ packages/samples/github_repository/lib/src/ui/github_repository_detail_page.dart
