+++ packages/samples/github_repository/lib/src/ui/provider/navigator_provider.dart
@@ -1,5 +1,5 @@
-import 'package:features_github_repository/src/ui/github_repository_list_page.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
+import 'package:samples_github_repository/src/ui/github_repository_list_page.dart';
 
 part 'navigator_provider.g.dart';
