+++ packages/samples/github_repository/pubspec.yaml
@@ -1,5 +1,5 @@
-name: features_github_repository
-description: "This package provides features for GitHub repository."
+name: samples_github_repository
+description: "This package provides sample GitHub repository functionality."
 publish_to: 'none'
 version: 0.0.1
